requirements.txt
streamlit
google-generativeai
python-dotenv
pydantic
pymupdf
